Add support for `--cddl-format` option
The cli currently has an intermediate TxBody type that is serialised in an intermediate format which does not follow cardano-ledger's CDDL spec.
A flag was added to the cli allowing the CDDL serialisation of any relevant outputs. Eventually the cli's intermediate format will be deprecated.

See https://github.com/input-output-hk/cardano-node/issues/3544